moved to build script
